# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification Service API",
        "description": "API documentation for the Notification Service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register-device": {
            "post": {
                "tags": [
                    "Device"
                ],
                "summary": "Register a device for push notifications",
                "operationId": "f8ef2c2d1c660dc44056e612f75b6289",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "user_id",
                                    "token",
                                    "platform"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "user_123"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "firebase_fcm_token_here"
                                    },
                                    "platform": {
                                        "type": "string",
                                        "example": "android"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Device registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device registered successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/device-token": {
            "get": {
                "tags": [
                    "Device"
                ],
                "summary": "Get or generate a device token",
                "operationId": "fb86f952b1e1919ab4abd092994be779",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "platform",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "web",
                                "android",
                                "ios"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token retrieved or generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "backend_generated_abc123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/send-notification": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Send a notification to specific users",
                "operationId": "bd6d9eb8a19fa0368ac2e48959ea1fdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title",
                                    "body",
                                    "user_ids"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "New Message"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "You have received a new message."
                                    },
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "user_123"
                                        }
                                    },
                                    "data": {
                                        "type": "object",
                                        "example": {
                                            "url": "/messages/1"
                                        },
                                        "additionalProperties": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification queued successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notification queued successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get all notifications for a user in a project",
                "operationId": "0bb334cc4f9efd4fc1e6a0b32d8cd65d",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "body": {
                                                "type": "string"
                                            },
                                            "data": {
                                                "type": "object",
                                                "nullable": true
                                            },
                                            "is_read": {
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mark-read/{id}": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Mark a notification as read",
                "operationId": "ae82c1a88783c43485390e7ed8382077",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notification marked as read"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Device",
            "description": "Device"
        },
        {
            "name": "Notification",
            "description": "Notification"
        }
    ]
}